feat: refactor StageRepository and update import paths
